Keep SignalActor and Semaphore in _common/test_utils.py so they remain accessible in the distributed Ray package
