+++ python/ray/_common/tests/test_utils.py
@@ -1,9 +1,15 @@
+"""Tests for Ray utility functions.
+
+This module contains pytest-based tests for utility functions in ray._common.utils.
+Test utility classes (SignalActor, Semaphore) are in ray._common.test_utils to
+ensure they're included in the Ray package distribution.
+"""
+
 import asyncio
 import warnings
 import sys
 
 import pytest
-import ray
 
 from ray._common.utils import (
     get_or_create_event_loop,
@@ -11,49 +17,10 @@ from ray._common.utils import (
     _BACKGROUND_TASKS,
 )
 
-# Export test utility classes for use by other test files
-__all__ = ["SignalActor", "Semaphore"]
 
-
-# Test utility classes (moved from _common/test_utils.py)
-@ray.remote(num_cpus=0)
-class SignalActor:
-    def __init__(self):
-        self.ready_event = asyncio.Event()
-        self.num_waiters = 0
-
-    def send(self, clear: bool = False):
-        self.ready_event.set()
-        if clear:
-            self.ready_event.clear()
-
-    async def wait(self, should_wait: bool = True):
-        if should_wait:
-            self.num_waiters += 1
-            await self.ready_event.wait()
-            self.num_waiters -= 1
-
-    async def cur_num_waiters(self) -> int:
-        return self.num_waiters
-
-
-@ray.remote(num_cpus=0)
-class Semaphore:
-    def __init__(self, value: int = 1):
-        self._sema = asyncio.Semaphore(value=value)
-
-    async def acquire(self):
-        await self._sema.acquire()
-
-    async def release(self):
-        self._sema.release()
-
-    async def locked(self) -> bool:
-        return self._sema.locked()
-
-
-# Tests for utility functions
 class TestGetOrCreateEventLoop:
+    """Tests for the get_or_create_event_loop utility function."""
+
     def test_existing_event_loop(self):
         # With running event loop
         expect_loop = asyncio.new_event_loop()
@@ -76,6 +43,7 @@ class TestGetOrCreateEventLoop:
 
 @pytest.mark.asyncio
 async def test_run_background_task():
+    """Test the run_background_task utility function."""
     result = {}
 
     async def co():
